+++ src/middleware/token.js
@@ -1,5 +1,19 @@
 export default function(context) {
-  // if no token, return
-  // validate token by calling database
-  console.log(context)
+  if (!context.query.token) {
+    return context.redirect('/')
+  }
+
+  const requestsCollection = context.app.$stitchDb.collection('requests')
+  const dbQuery = {
+    token: context.query.token,
+  }
+
+  requestsCollection
+    .findOne(dbQuery)
+    .then((result) => {
+      if (!result) {
+        return context.redirect('/')
+      }
+    })
+    .catch((err) => console.error(err))
 }