+++ .github/CHANGELOG.md
@@ -7,3 +7,6 @@
 - installed `node-sass` for Sass styling integration
 - added `dotenv` and configured development server to load setup from `.env`
 - added `mongodb-stitch-browser-sdk`
+- added `mongodb-stitch-browser-services-aws`
+- added `filepond` and `vue-filepond`
+- created file upload handler on `submit` page